FrontendUI/requirements.txt
python-dotenv==0.20.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.1
Flask-WTF==1.0.1
WTForms==3.0.1
requests==2.28.1
Werkzeug==2.1.2
pip==22.1.2
setuptools==62.6.0
blinker==1.4
wheel==0.37.1
cryptography==37.0.4





